requirements.txt
ultralytics==8.2.4
pandas==2.0.2
opencv-python==4.9.0.80
numpy==1.24.3
scipy==1.10.1
filterpy==1.4.5
deep_sort_realtime==1.3.2
pillow==10.3.0
